requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.49
psycopg2-binary
Werkzeug
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0
